@@ -3,38 +3,84 @@
   <PropertyGroup>
     <TargetFramework>net472</TargetFramework>
     <AllowUnsafeBlocks>true</AllowUnsafeBlocks>
-    <LangVersion>9.0</LangVersion>
+    <LangVersion>10.0</LangVersion>
     <GenerateAssemblyInfo>false</GenerateAssemblyInfo>
-    <OutputPath>..\..\Assemblies\</OutputPath>
+    <OutputPath Condition=" '$(Configuration.EndsWith( 1.3))' ">..\..\1.3\Assemblies\</OutputPath>
+    <OutputPath Condition=" '$(Configuration.EndsWith( 1.4))' ">..\..\1.4\Assemblies\</OutputPath>
     <AppendTargetFrameworkToOutputPath>false</AppendTargetFrameworkToOutputPath>
     <AppendRuntimeIdentifierToOutputPath>false</AppendRuntimeIdentifierToOutputPath>
     <DebugType>portable</DebugType>
     <PublicizeAsReferenceAssemblies>false</PublicizeAsReferenceAssemblies>
     <RootNamespace>$(MSBuildProjectName.Replace(" ", "."))</RootNamespace>
+    <Configurations>Release 1.3;Debug 1.4;Debug 1.3;Release 1.4</Configurations>
+    <Platforms>AnyCPU</Platforms>
+  </PropertyGroup>
+
+  <PropertyGroup Condition=" '$(Configuration.StartsWith(Release))' ">
+    <Optimize Condition=" '$(Optimize)' == '' ">true</Optimize>
+    <DefineConstants>TRACE;RELEASE;</DefineConstants>
+  </PropertyGroup>
+  <PropertyGroup Condition=" '$(Configuration.StartsWith(Debug))' ">
+    <DebugSymbols Condition=" '$(DebugSymbols)' == '' ">true</DebugSymbols>
+    <Optimize Condition=" '$(Optimize)' == '' ">false</Optimize>
+    <DefineConstants>TRACE;DEBUG;</DefineConstants>
+  </PropertyGroup>
+  <PropertyGroup Condition=" '$(Configuration.EndsWith( 1.3))' ">
+    <DefineConstants>IDEOLOGY;IDEOLOGY_PLUS</DefineConstants>
+  </PropertyGroup>
+  <PropertyGroup Condition=" '$(Configuration.EndsWith( 1.4))' ">
+    <DefineConstants>BIOTECH;BIOTECH_PLUS;IDEOLOGY_PLUS</DefineConstants>
   </PropertyGroup>
 
   <ItemGroup>
     <PackageReference Include="Krafs.Publicizer" Version="1.0.*" />
     <PackageReference Include="Lib.Harmony" Version="2.2.*" ExcludeAssets="runtime" />
-    <PackageReference Include="Krafs.Rimworld.Ref" Version="1.3.*" />
+    <PackageReference Condition=" '$(Configuration.EndsWith( 1.3))' " Include="Krafs.Rimworld.Ref" Version="1.3.*" />
+    <PackageReference Condition=" '$(Configuration.EndsWith( 1.4))' " Include="Krafs.Rimworld.Ref" Version="1.4.*" />
   </ItemGroup>
 
   <ItemGroup>
     <None Remove=".editorconfig" />
     <None Remove="mono_crash.*.json" />
     <None Include="..\..\About\**" LinkBase="Common\About" />
-    <None Include="..\..\Defs\**" LinkBase="Common\Defs" />
+    <None Include="..\..\1.3\Defs\**" LinkBase="Common 1.3\Defs" />
+    <None Include="..\..\1.4\Defs\**" LinkBase="Common 1.4\Defs" />
     <None Include="..\..\Languages\**" LinkBase="Common\Languages" />
-    <None Include="..\..\Patches\**" LinkBase="Common\Patches" />
+    <None Include="..\..\1.3\Patches\**" LinkBase="Common 1.3\Patches" />
+    <None Include="..\..\1.4\Patches\**" LinkBase="Common 1.4\Patches" />
   </ItemGroup>
 
   <ItemGroup>
     <Publicize Include="Assembly-CSharp" />
+    <Publicize Include="VFEAncients" />
   </ItemGroup>
 
-  <ItemGroup>
+  <ItemGroup Condition=" '$(Configuration.EndsWith( 1.3))' ">
     <Reference Include="Rimefeller">
-      <HintPath>..\..\References\Rimefeller.dll</HintPath>
+      <HintPath>..\..\..\..\..\..\workshop\content\294100\1321849735\1.3\Assemblies\Rimefeller.dll</HintPath>
+      <Private>false</Private>
+    </Reference>
+    <Reference Include="VFEAncients">
+      <HintPath>..\..\..\..\..\..\workshop\content\294100\2654846754\1.3\Assemblies\VFEAncients.dll</HintPath>
+      <Private>false</Private>
+    </Reference>
+    <Reference Include="VanillaPowerExpanded">
+      <HintPath>..\..\..\..\..\..\workshop\content\294100\2062943477\1.3\Assemblies\VanillaPowerExpanded.dll</HintPath>
+      <Private>false</Private>
+    </Reference>
+  </ItemGroup>
+
+  <ItemGroup Condition=" '$(Configuration.EndsWith( 1.4))' ">
+    <Reference Include="Rimefeller">
+      <HintPath>..\..\..\..\..\..\workshop\content\294100\1321849735\1.4\Assemblies\Rimefeller.dll</HintPath>
+      <Private>false</Private>
+    </Reference>
+    <Reference Include="VFEAncients">
+      <HintPath>..\..\..\..\..\..\workshop\content\294100\2654846754\1.4\Assemblies\VFEAncients.dll</HintPath>
+      <Private>false</Private>
+    </Reference>
+    <Reference Include="VanillaPowerExpanded">
+      <HintPath>..\..\..\..\..\..\workshop\content\294100\2062943477\1.4\Assemblies\VanillaPowerExpanded.dll</HintPath>
       <Private>false</Private>
     </Reference>
   </ItemGroup>
