@@ -1,30 +1,44 @@
 ﻿using System;
 using Verse;
 
-namespace MoreRimefellerCompat
-{
-    public class JunctionToPipeNetComp : ThingComp
-    {
-        private IJunctionToPipeNet junctionToPipeNetComp;
-        public static Func<JunctionToPipeNetComp, IJunctionToPipeNet> OnPostSpawnSetup;
+namespace MoreRimefellerCompat;
 
-        public override void PostSpawnSetup(bool respawningAfterLoad)
+public class JunctionToPipeNetComp : ThingComp
+{
+    private IJunctionToPipeNet junctionToPipeNetComp;
+    public static Func<JunctionToPipeNetComp, IJunctionToPipeNet> onPostSpawnSetup = _ => new DummyImplementation();
+
+    public override void PostSpawnSetup(bool respawningAfterLoad)
+    {
+        junctionToPipeNetComp = onPostSpawnSetup?.Invoke(this);
+        junctionToPipeNetComp?.PostSpawnSetup(respawningAfterLoad);
+    }
+
+    public override void ReceiveCompSignal(string signal) => junctionToPipeNetComp?.ReceiveCompSignal(signal);
+
+    public override void CompTick() => junctionToPipeNetComp?.CompTick();
+
+    public interface IJunctionToPipeNet
+    {
+        void PostSpawnSetup(bool respawningAfterLoad);
+
+        void ReceiveCompSignal(string signal);
+
+        void CompTick();
+    }
+        
+    private class DummyImplementation : IJunctionToPipeNet
+    {
+        public void PostSpawnSetup(bool respawningAfterLoad)
         {
-            junctionToPipeNetComp = OnPostSpawnSetup?.Invoke(this);
-            junctionToPipeNetComp?.PostSpawnSetup(respawningAfterLoad);
         }
 
-        public override void ReceiveCompSignal(string signal) => junctionToPipeNetComp?.ReceiveCompSignal(signal);
-
-        public override void CompTick() => junctionToPipeNetComp?.CompTick();
-
-        public interface IJunctionToPipeNet
+        public void ReceiveCompSignal(string signal)
         {
-            void PostSpawnSetup(bool respawningAfterLoad);
+        }
 
-            void ReceiveCompSignal(string signal);
-
-            void CompTick();
+        public void CompTick()
+        {
         }
     }
 }
